workers/vision/requirements.txt
# AI Vision Worker Dependencies
# Virtual Drone Simulator: YOLOv8 + CLIP + NATS + Datalastic AIS
ultralytics>=8.1.0
nats-py
opencv-python
numpy
httpx
torch
torchvision
transformers
Pillow
pymavlink==2.4.41
nats-py==2.6.0